refactor: refactor the dto type files
