+++ types/DestinationDto.ts
@@ -1,0 +1,7 @@
+export type DestinationDto = {
+    id: number
+    name: string
+    //NEED TO FIND THE RIGHT TYPE FOR COORDINATES
+    coordinates: string
+  }
+
