+++ types/RestaurantDto.ts
@@ -1,0 +1,5 @@
+export type RestaurantDto = {
+    id: number
+    name: string
+    coordinates: string
+  }
